backend/requirements.txt
# Core Flask dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.3
python-dotenv>=1.0.0
werkzeug>=3.0.1

# Scientific computing (use newer versions with pre-built wheels)
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Deep learning (install separately if needed - large files)
# ultralytics>=8.0.0
# torch>=2.0.0
# torchvision>=0.15.0

# Server
gunicorn>=21.0.0

# Database drivers (install based on your database)
# For MySQL/MariaDB (XAMPP):
pymysql>=1.1.0
# OR mysql-connector-python>=8.2.0

# For PostgreSQL (optional):
# psycopg2-binary>=2.9.9

